Split out frontmatter getters
