+++ src/frontmatter/coverImage.js
@@ -1,0 +1,3 @@
+module.exports = (post) => {
+  return post.meta.coverImage;
+};
