+++ src/frontmatter/title.js
@@ -1,0 +1,3 @@
+module.exports = (post) => {
+  return post.data.title[0];
+};
